+++ ChangeLog.rst
@@ -9,6 +9,7 @@
 * Made a top level timeout for 16 minutes on a single project check. During this window,
   Ultralisp will make 3 attempts and each will be limited by 5 minutes timeout. If some
   project takes longer for compile or hangs, it will fail.
+* Page of all tags now uses cached results.
 
 1.16.7 (2022-09-21)
 ===================